refactor!(core): Properly move distributed code from examples to core
