+++ core/CMakeLists.txt
@@ -23,6 +23,17 @@ set(SOURCE_FILES
     src/cpu/adapter_cblas_fp32.cpp
     src/cpu/adapter_cblas_fp64.cpp)
 
+# GPRat distributed TODO: this could be gated behind a distributed-only option!
+list(
+  APPEND
+  SOURCE_FILES
+  src/cpu/adapter_cblas_fp64_actions.cpp
+  src/cpu/gp_algorithms_actions.cpp
+  src/cpu/gp_uncertainty_actions.cpp
+  src/cpu/gp_optimizer_actions.cpp
+  src/tile_cache.cpp
+  src/tiled_dataset.cpp)
+
 if(GPRAT_WITH_CUDA)
   list(
     APPEND
