+++ core/src/tiled_dataset.cpp
@@ -1,0 +1,3 @@
+﻿#include "gprat/tiled_dataset.hpp"
+
+GPRAT_REGISTER_TILED_DATASET(double, double);
