--- examples/distributed/src/scheduling.cpp
@@ -1,2 +1,0 @@
-﻿#include "scheduling.hpp"
-#include "distributed_tile.hpp"
