--- examples/distributed/src/scheduling.hpp
@@ -1,10 +1,0 @@
-﻿#pragma once
-
-#include "gprat/detail/config.hpp"
-
-#include "gprat/detail/async_helpers.hpp"
-#include "gprat/detail/actions.hpp"
-
-GPRAT_NS_BEGIN
-
-GPRAT_NS_END
